Move CameraView to ui
