--- firebase-ai/app/src/main/java/com/google/firebase/quickstart/ai/feature/media/imagen/EditingMode.kt
@@ -1,10 +1,0 @@
-package com.google.firebase.quickstart.ai.feature.media.imagen
-
-enum class EditingMode {
-    GENERATE,
-    INPAINTING,
-    OUTPAINTING,
-    SUBJECT_REFERENCE,
-    STYLE_TRANSFER,
-    TEMPLATE,
-}
